moved uatraits to COmmon, fixing mistakes
