+++ programs/server/Server.cpp
@@ -31,7 +31,7 @@
 #include <Common/getExecutablePath.h>
 #include <Common/ThreadProfileEvents.h>
 #include <Common/ThreadStatus.h>
-#include <Core/UatraitsFast/UserAgent.h>
+#include <Common/UatraitsFast/UserAgent.h>
 #include <IO/HTTPCommon.h>
 #include <IO/UseSSL.h>
 #include <Interpreters/AsynchronousMetrics.h>
