+++ src/Common/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_subdirectory(StringUtils)
+add_subdirectory(UatraitsFast)
 # after common_io
 #add_subdirectory(ZooKeeper)
 #add_subdirectory(ConfigProcessor)
