+++ src/Common/UatraitsFast/CMakeLists.txt
@@ -1,3 +1,2 @@
 add_subdirectory(types)
 add_subdirectory(data)
-add_subdirectory(uatraits)
