+++ src/Common/UatraitsFast/UATraitsCaseSafeCached.h
