+++ src/Common/UatraitsFast/types/OperatingSystem.cpp
