+++ src/Core/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory (UatraitsFast)
 if (ENABLE_TESTS)
     add_subdirectory (tests)
 endif ()
